Moved RooRealL plotting test to roofitcore/test
